% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ships_Final.R
\docType{data}
\name{Ships_Final}
\alias{Ships_Final}
\title{A data frame with longest sailed distance for each ship}
\format{
A data frame with 1208 rows and 29 variables.
}
\usage{
Ships_Final
}
\description{
A data frame with longest sailed distance for each ship
}
\keyword{datasets}
